Filename cleanup before deployment
