+++ nuxt-app/components/SearchResultCard.vue
